submit java-oop-ru/errors

+++ java-oop-ru/errors/src/main/java/exercise/App.java
@@ -4,7 +4,7 @@ package exercise;
 class App {
     public static void printSquare(Circle circle) {
         try {
-            System.out.println(Math.ceil(circle.getSquare()));
+            System.out.println((int) Math.ceil(circle.getSquare()));
         } catch (NegativeRadiusException e) {
             System.out.println("Не удалось посчитать площадь");
         } finally {
